イベント一覧画面一通りV0.9分を実装完了。

+++ src/jp/vocalendar/activity/EventLoadingActivity.java
@@ -1,7 +1,5 @@
 package jp.vocalendar.activity;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.Date;
@@ -17,6 +15,7 @@ import jp.vocalendar.googleapi.OAuthManager;
 import jp.vocalendar.model.Event;
 import jp.vocalendar.model.EventComparator;
 import jp.vocalendar.model.EventDataBase;
+import jp.vocalendar.model.EventSeparator;
 import jp.vocalendar.model.GoogleCalendarLoadEventTask;
 import jp.vocalendar.model.LoadEventTask;
 import android.accounts.Account;
@@ -30,13 +29,12 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import edu.emory.mathcs.backport.java.util.LinkedList;
 
 /**
- * スプラッシュ画面を表示するActivity。
+ * イベント読み込み中の画面のActivity。
  * バックグランド処理でイベントを読み込む。
  */
-public class SplashScreenActivity extends Activity implements LoadEventTask.TaskCallback {
+public class EventLoadingActivity extends Activity implements LoadEventTask.TaskCallback {
 	private static String TAG = "SplashScreenActivity";
 	
 	/**
@@ -47,8 +45,6 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 	private GoogleCalendarLoadEventTask task = null;	
 	private TextView loadingItemView = null;
 	
-	private List<Event> foundEvents = new LinkedList();
-		
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -72,25 +68,39 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 
 	private void startLoadEventTask() {		
         TimeZone timeZone = TimeZone.getDefault();
-        java.util.Calendar cal = java.util.Calendar.getInstance(timeZone); // いったんローカルのタイムゾーンでカレンダー計算
-		cal.set(java.util.Calendar.HOUR_OF_DAY, 0);
-		cal.set(java.util.Calendar.MINUTE, 0);
-		cal.set(java.util.Calendar.SECOND, 0);
-		cal.set(java.util.Calendar.MILLISECOND, 0);
-		cal.setTimeZone(TimeZone.getTimeZone("UTC")); // ローカルのタイムゾーンで日付設定したら、タイムゾーンをUTCに強制変更
-		DateTime start = new DateTime(cal.getTimeInMillis(), 0); // tzShiftを0。UTCのオフセットを"Z"(=00:00)にする(参考：RFC3339)
-		cal.add(java.util.Calendar.DATE, +1);
-		DateTime end = new DateTime(cal.getTimeInMillis(), 0);
-
+        Calendar localCal = Calendar.getInstance(timeZone); // いったんローカルのタイムゾーンでカレンダー計算
+		
+        int year = localCal.get(Calendar.YEAR);
+        int month = localCal.get(Calendar.MONTH);
+        int date = localCal.get(Calendar.DATE);
+        
+        Calendar utcCal = Calendar.getInstance(TimeZone.getTimeZone("UTC")); // タイムゾーンUTCで、ローカルのタイムゾーンの日付を指定
+        utcCal.set(Calendar.YEAR, year);
+        utcCal.set(Calendar.MONTH, month);
+        utcCal.set(Calendar.DATE, date);        
+        utcCal.set(Calendar.HOUR_OF_DAY, 0);
+        utcCal.set(Calendar.MINUTE, 0);
+        utcCal.set(Calendar.SECOND, 0);
+        utcCal.set(Calendar.MILLISECOND, 0);
+                
+        DateTime[] dates = new DateTime[getNumberOfDateToGetEvent() + 1];
+        EventSeparator[] separators = new EventSeparator[dates.length - 1];
+    	dates[0] = new DateTime(utcCal.getTimeInMillis(), 0); // tzShiftを0。UTCのオフセットを"Z"(=00:00)にする(参考：RFC3339)
+        for(int i = 0; i < separators.length; i++) {
+        	separators[i] = new EventSeparator(localCal.getTime());
+        	localCal.add(Calendar.DATE, 1);
+        	
+    		utcCal.add(java.util.Calendar.DATE, 1);
+    		dates[i+1] = new DateTime(utcCal.getTimeInMillis(), 0);        	
+        }
+        
 		task = new GoogleCalendarLoadEventTask(this, this, 5);
-		task.setStartAndEndDate(start, end, timeZone);
+		task.setStartAndEndDate(dates, separators, timeZone);
 		task.execute(Constants.MAIN_CALENDAR_ID, Constants.BROADCAST_CALENDAR_ID);
 	}
 	
 	public void transitToEventListActivity() {
-		Intent i = new Intent(this, EventListActivity.class);
-		i.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-		startActivity(i);		
+		setResult(RESULT_OK);
 		finish();
 	}
 
@@ -102,22 +112,22 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 		String str = event.toDateTimeSummaryString();
 		Log.d("SplashScreenActivity", str);
 		loadingItemView.setText(str);
-		//if(event.equalByDate(today, timeZone)) {
-		foundEvents.add(event);
-		//}
 	}
 	
-	public void onPostExecute() {		
+	public void onPostExecute(List<Event> events) {
+		if(events == null) {
+			return;
+		}		
 		loadingItemView.setText("表示準備中");
-		
-		Collections.sort(foundEvents, new EventComparator()); 
 		
 		EventDataBase db = new EventDataBase(this);
 		db.open();
 		db.deleteAllEvent();
-		Iterator<Event> itr = foundEvents.iterator();
+				
+		Iterator<Event> itr = events.iterator();
+		int index = 0;
 		while(itr.hasNext()) {
-			db.insertEvent(itr.next());
+			db.insertEvent(index++, itr.next());
 		}
 		db.close();		
 		
@@ -129,9 +139,7 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 	 */
 	public void retry(int tryNumber) {		
 		Log.d(TAG, "retry()");
-		foundEvents.clear();		
-		new GoogleCalendarLoadEventTask(this, this, tryNumber).execute(
-				Constants.MAIN_CALENDAR_ID, Constants.BROADCAST_CALENDAR_ID);		
+		startLoadEventTask();
 	}
 	
 	
@@ -150,7 +158,6 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 			showGoogleAccountRequredDialog();
 			return;
 		}
-		
 		Log.i(TAG, "onAuthToken: " + account.name + "," + authToken);
 		startLoadEventTask();
 	}	
@@ -183,5 +190,13 @@ public class SplashScreenActivity extends Activity implements LoadEventTask.Task
 		if(requestCode == LOGIN_REQUEST_CODE) {
 			initAccount();
 		}
+	}
+	
+	/**
+	 * イベント取得日数を返す。
+	 * @return
+	 */
+	private int getNumberOfDateToGetEvent() {
+		return 3;
 	}
 }
